feat: create the core of sprout
