+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/42atomys/sprout
 
-go 1.19
+go 1.21
 
 require (
 	dario.cat/mergo v1.0.0
